Fix: Update README.md and relocate test files to the appropriate directories.
